Last of the game DLL level 4 warning fixes.

+++ game/g_svcmds.c
@@ -100,7 +100,7 @@ static qboolean StringToFilter (char *s, ipfilter_t *f)
 			num[j++] = *s++;
 		}
 		num[j] = 0;
-		b[i] = atoi(num);
+		b[i] = (byte)atoi(num);
 		if (b[i] != 0)
 			m[i] = 255;
